+++ src/main/java/com/rocketpt/server/dto/sys/MenuResourceDTO.java
@@ -3,6 +3,6 @@ package com.rocketpt.server.dto.sys;
 /**
  * @author plexpt
  */
-public record MenuResourceDTO(Long id, String name, String url, String icon, Long parentId) {
+public record MenuResourceDTO(Integer id, String name, String url, String icon, Integer parentId) {
 
 }
